requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pypdf
openai
pillow
python-dotenv